copy updates from https://github.com/rbdannenberg/portmidi
The original author of PortMidi, Roger B. Dannenberg recently made
some updates to the C code, including incorporating changes found
in various PortMidi forks.

The organization of that Git repository is different so the commits
cannot simply be cherry picked or rebased. I manually copied the
updated C code files to the places they have been moved to in this
repository.

Functional changes:
* added Pm_CreateVirtualInput() and Pm_CreateVirtualOutput()
  functions that allow applications to create named ports
  analogous to devices.
* improvements for macOS CoreMIDI include higher data rates for
  devices, better handling of Unicode interface names in addition
  to virtual device creation.
* the notion of default devices, Pm_GetDefaultInputDeviceID(),
  Pm_GetDefaultOutputDeviceID and the PmDefaults program have
  fallen into disuse and are now deprecated.